+++ engineFoxtrot/Engine/Structs/fpsStructs.h
@@ -5,10 +5,10 @@ class DLLEXPORT Message
 struct FpsMessage
 #endif
 {
-	FpsMessage(const string& _text, int _r, int _g, int _b)
+	FpsMessage(const string& _text, uint8_t _r, uint8_t _g, uint8_t _b)
 		: text{ _text }, red{ _r }, green{ _g }, blue{ _b }{}
 	const string& text;
-	int red, green, blue;
+	uint8_t red, green, blue;
 };
 
 #if(EXPORT)
